ai_backend/requirements.txt
fastapi
pydantic
supabase
google-generativeai
uvicorn
